Split code into modules: units, road parts, road functions and networks
